Crate linera_version

This crate is in charge of extracting version information from the Linera build, for troubleshooting information and version compatibility checks.

Structs

ApiHashes

CrateVersion

Pretty

VersionInfo

The version info of a build of Linera.

Enums

Error

Statics

VERSION_INFO

Type Aliases

Hash